requirements.txt
git+https://github.com/cibere/kick.py
spotipy
python-dotenv